+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [Block Types](https://developer.wordpress.org/rest-api/reference/block-types/) endpoint
+
 ## [0.2.0] - 2026-05-05
 
 ### Added